pyproject.toml
[project]
name = "rag-benchmarking"
version = "0.1.0"
description = "Docling + LangChain RAG agent with Qdrant and HuggingFace"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain==1.2.10",
    "langchain-community==0.4.1",
    "langchain-core==1.2.18",
    "langchain-docling==2.0.0",
    "langchain-huggingface==1.2.1",
    "langchain-text-splitters==1.1.1",
    "langchain-ollama==1.0.1",
    "qdrant-client==1.17.0",
    "fastembed==0.7.4",
    "python-dotenv==1.2.2",
    "sentence-transformers==5.2.3",
    "transformers==4.57.6",
]
